+++ scripts/autorun/libs/utils.lua
@@ -10,6 +10,17 @@ function table.count(tab)
 	return size
 end
 
+function table.join(...)
+	local t = {}
+	for _, tab in pairs({...}) do
+		for k, v in ipairs(tab) do
+			table.insert(t, v)
+		end
+	end
+
+	return t
+end
+
 -- Utils table
 utils = {}
 function utils.OverrideFunction(originalFunc, newFunc)